Extract common eth1 functionality to separate module
